+++ tests/cli-main.test.ts
@@ -43,6 +43,10 @@ class TestUi implements AgentifyUi {
     throw new Error("promptMultiSelect should not be called in this test");
   }
 
+  async promptCheckboxList(): Promise<ReadonlyArray<string>> {
+    throw new Error("promptCheckboxList should not be called in this test");
+  }
+
   async promptSecret(): Promise<string> {
     throw new Error("promptSecret should not be called in this test");
   }